feat: new script
sd2ba_fetch-data.py is a script that fetches the AA and nucleotide sequences from the proteins present
in a UniRef cluster. The input is a UniRef cluster ID and the output is a fasta
file with the AA sequences and another fasta file with the nucleotide sequences.

sd2ba_fetch-data.py will fetch the data and sd2ba.py will do the actual analysis.