feat: HMMER call implemented
Now the script runs HMMER to perfom an aligment with HMMALIGN. HMMALIGN
returns a file in afa format so the stdout is piped to sed in order to
replace al "." with "-", resulting in a fasta file as output.

TODO: Make all of this into a function so the main function is not
overloaded.